+++ packages/grill/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @macalinao/grill
 
+## 0.7.2
+
+### Patch Changes
+
+- c9b6849: Update dependencies
+- 3f9e7d9: Remove unnecessary arguments
+- c093c0e: Fix broken token metadata fetching -- allow for empty symbols
+- 8964b7b: Update dependencies
+- Updated dependencies [c9b6849]
+- Updated dependencies [3f9e7d9]
+- Updated dependencies [c093c0e]
+- Updated dependencies [0a8bc95]
+- Updated dependencies [8964b7b]
+  - @macalinao/token-utils@0.1.11
+  - @macalinao/gill-extra@0.2.1
+
+## 0.7.1
+
+### Patch Changes
+
+- 90affa8: Add ALT fetching hooks
+- f4b610c: Reorganize directory structure of account/pda fetchers
+
 ## 0.7.0
 
 ### Minor Changes